Issue-133: Created IEntertainmentMapper and it's implementation EntertainmentMapper
