+++ app/src/main/java/com/kirchhoff/movies/di/MapperModule.kt
@@ -2,6 +2,8 @@ package com.kirchhoff.movies.di
 
 import com.kirchhoff.movies.mapper.discover.DiscoverMapper
 import com.kirchhoff.movies.mapper.discover.IDiscoverMapper
+import com.kirchhoff.movies.mapper.entertainment.EntertainmentMapper
+import com.kirchhoff.movies.mapper.entertainment.IEntertainmentMapper
 import com.kirchhoff.movies.mapper.movie.IMovieDetailsMapper
 import com.kirchhoff.movies.mapper.movie.MovieDetailsMapper
 import com.kirchhoff.movies.mapper.person.details.IPersonDetailsMapper
@@ -15,9 +17,10 @@ import com.kirchhoff.movies.mapper.tv.TvDetailsMapper
 import org.koin.dsl.module
 
 val mapperModule = module {
+    single<IEntertainmentMapper> { EntertainmentMapper() }
     single<IPersonDetailsMapper> { PersonDetailsMapper() }
-    single<IMovieDetailsMapper> { MovieDetailsMapper() }
-    single<ITvDetailsMapper> { TvDetailsMapper() }
+    single<IMovieDetailsMapper> { MovieDetailsMapper(entertainmentMapper = get()) }
+    single<ITvDetailsMapper> { TvDetailsMapper(entertainmentMapper = get()) }
     single<IReviewListMapper> { ReviewListMapper() }
     single<IDiscoverMapper> { DiscoverMapper() }
     single<IPersonsMapper> { PersonsMapper() }
