+++ app/src/main/java/com/kirchhoff/movies/mapper/tv/TvDetailsMapper.kt
@@ -1,17 +1,16 @@
 package com.kirchhoff.movies.mapper.tv
 
 import com.kirchhoff.movies.data.network.core.NetworkEntertainmentCredits
-import com.kirchhoff.movies.data.network.core.NetworkEntertainmentPerson
 import com.kirchhoff.movies.data.network.core.NetworkObjectWithName
 import com.kirchhoff.movies.data.network.details.tv.NetworkTvDetails
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
-import com.kirchhoff.movies.data.ui.core.UIEntertainmentPerson
 import com.kirchhoff.movies.data.ui.core.UIGenre
 import com.kirchhoff.movies.data.ui.details.tv.UITvDetails
 import com.kirchhoff.movies.mapper.BaseMapper
+import com.kirchhoff.movies.mapper.entertainment.IEntertainmentMapper
 import com.kirchhoff.movies.repository.Result
 
-class TvDetailsMapper : BaseMapper(), ITvDetailsMapper {
+class TvDetailsMapper(private val entertainmentMapper: IEntertainmentMapper) : BaseMapper(), ITvDetailsMapper {
 
     override fun createUITvDetails(tvDetailsResult: Result<NetworkTvDetails>): Result<UITvDetails> =
         when (tvDetailsResult) {
@@ -21,7 +20,7 @@ class TvDetailsMapper : BaseMapper(), ITvDetailsMapper {
 
     override fun createUIEntertainmentCredits(tvCreditsResult: Result<NetworkEntertainmentCredits>): Result<UIEntertainmentCredits> =
         when (tvCreditsResult) {
-            is Result.Success -> Result.Success(createUITvCredits(tvCreditsResult.data))
+            is Result.Success -> Result.Success(entertainmentMapper.createUIEntertainmentCredits(tvCreditsResult.data))
             else -> mapErrorOrException(tvCreditsResult)
         }
 
@@ -35,28 +34,6 @@ class TvDetailsMapper : BaseMapper(), ITvDetailsMapper {
             tvDetails.vote_count,
             tvDetails.vote_average,
             tvDetails.genres.map { createUIGenre(it) }
-        )
-
-    private fun createUITvCredits(tvCredits: NetworkEntertainmentCredits) =
-        UIEntertainmentCredits(
-            tvCredits.cast?.map { createUIEntertainmentActor(it) },
-            tvCredits.crew?.map { createUIEntertainmentCreator(it) }
-        )
-
-    private fun createUIEntertainmentActor(actor: NetworkEntertainmentPerson.Actor) =
-        UIEntertainmentPerson.Actor(
-            actor.id,
-            actor.name,
-            actor.profile_path,
-            actor.character
-        )
-
-    private fun createUIEntertainmentCreator(creator: NetworkEntertainmentPerson.Creator) =
-        UIEntertainmentPerson.Creator(
-            creator.id,
-            creator.name,
-            creator.profile_path,
-            creator.job
         )
 
     private fun createUIGenre(item: NetworkObjectWithName) =
